Step 42: Cook 2019 connectome calibration + perf optimization + fitness framework
- Connectome: AVD→AVA 1→2, AIB→AVA 3/3→2/1, AIY→RIA 4→5 (Cook 2019 EM)
- Delete AVE→RIV (not in Cook 2019, caused RIV tonic activation)
- Add RIA↔RIV negative feedback loop: RIA→RIV(ACh) + RIV→RIA(GABA) + RIV↔RIV gap
- Params: pulse_amp=50, as_factor=1.7 (loop compensation)
- Perf P0: cache awc_pref (eliminate 300M string ops/300s)
- Perf P0: cache 10 neuron IDs (eliminate 10.8M hash lookups/300s)
- Perf P1: cache 6 dynamic_cast pointers (eliminate 3.6M RTTI/300s)
- Perf P1: pre-index AWC/ASER learning synapses (eliminate full scan)
- Fitness: --fitness CLI, SimMetrics, 4 seeds x 3 scenarios auto-eval
- Docs: step42_connectome_calibration.md + progress.md Step 41/42
- regtest: 17 pass, 0 FAIL

+++ src/core/types.h
@@ -86,7 +86,7 @@ struct NeuronInfo {
 struct SynapseInfo {
     int pre_neuron_id = -1;
     int post_neuron_id = -1;
-    int num_sections = 1;     // EM section count (anatomical strength proxy)
+    double num_sections = 1.0;  // EM section count (double: supports extrasynaptic <1.0)
     NeurotransmitterType neurotransmitter = NeurotransmitterType::UNKNOWN;
     int post_compartment = 0; // Step 28: target compartment (0=soma, default)
 };
@@ -94,7 +94,7 @@ struct SynapseInfo {
 struct GapJunctionInfo {
     int neuron_a_id = -1;
     int neuron_b_id = -1;
-    int num_sections = 1;     // EM section count
+    double num_sections = 1.0;  // EM section count (double: supports fractional)
     int compartment_a = 0;    // Step 28: compartment on neuron A (0=soma)
     int compartment_b = 0;    // Step 28: compartment on neuron B (0=soma)
 };
